requirements.txt
groq>=0.4.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.3.0
scikit-learn>=1.3.0
faker>=22.0.0
streamlit>=1.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
